Moves the getNumberOfLinesFromUser function to a new module called inputs
